move multiprocessing module out of profiler module so other module could use it
